+++ packages/http-client/src/lib/clients/alerts/AlertsClient.types.ts
@@ -150,3 +150,13 @@ export interface ClientConfig {
   baseUrl?: string
   accessToken?: string
 }
+
+export type PaginatedListEntity<T> = {
+  total: number | null
+  page: number
+  next: number | null
+  hasNext: boolean
+  prev: number | null
+  hasPrev: boolean
+  items: T[]
+}