+++ packages/http-client/src/lib/clients/stamps/__examples__/StampsClient.example.ts
@@ -1,10 +1,13 @@
 import { isHttpError, isParseError, StampsClient } from '../../../..'
 import { Avro, GetAvroParams } from '../StampsClient.types'
+import { Blob } from 'buffer'
 
 export async function exampleGetAvroJson(): Promise<Avro> {
   const params: GetAvroParams = {
     candid: '1450198835415015001',
     survey_id: 'ztf',
+    type: 'science',
+    format: 'png',
   }
   try {
     const result = await StampsClient.getAvroJson<Avro>(params)
@@ -31,6 +34,44 @@ export async function exampleGetAvroJson(): Promise<Avro> {
       // console.error(exception)
       throw exception
     }
+    throw exception
+  }
+}
+
+export async function exampleGetStamp(): Promise<Blob> {
+  const params: GetAvroParams = {
+    candid: '1450198835415015001',
+    survey_id: 'ztf',
+    type: 'science',
+    format: 'png',
+  }
+
+  try {
+    const result = await StampsClient.getStamp<Blob>(params)
+    return result
+  } catch (exception) {
+    throw exception
+  }
+}
+
+export async function exampleImportingBase64Parser(): Promise<Promise<string>> {
+  const utils = await import('../utils')
+  const params: GetAvroParams = {
+    candid: '1450198835415015001',
+    survey_id: 'ztf',
+    type: 'science',
+    format: 'png',
+  }
+
+  try {
+    const result = await StampsClient.getStamp<Promise<string>>(
+      params,
+      utils.b64Parser
+    )
+    console.log('example', result)
+    return result
+  } catch (exception) {
+    console.log('example error', exception)
     throw exception
   }
 }