+++ packages/http-client/src/lib/util.types.ts
@@ -1,11 +1,2 @@
 export type Newable<T> = { new (...args: unknown[]): T }
 
-export type PaginatedListEntity<T> = {
-  total: number | null
-  page: number
-  next: number | null
-  hasNext: boolean
-  prev: number | null
-  hasPrev: boolean
-  items: T[]
-}